Apply less zoom by default
Actually, the current default zoom level is fine for the linear scale.
But since the logarithmic scale is enabled by default, the default zoom
level should be adapted to the logarithmic scale.

A proper fix would be to have a comparable size for both types of scale,
but that may prove difficult to implement.